Fix for disk speed problem
The original code was accumulating disks/block devices and partitions, for read and write sectors. Now the code gets the list of block devices and just accumulate its sectors.
I found a division of this sectors of 512 bytes by 1024 and 8 to get MiB, but this is not right, it is needed to divide for 2 not 8, maybe because of accumulating block devices and partitions.
Nowadays, disk space is in MB, not in MiB, so I adjust this too.