README: Overhaul!
Overhaul the README to make it more digestible and better organized. This
will help users troubleshoot common issues (e.g. "Meta is null") and is
just good marketing.

It would be great to fix the dependency check message pop-up (tracked in
issue #69, nice), but we haven't done so yet out of sheer laziness. Adding
the "Meta is null" error message to the README should help a bit with user
self-troubleshooting.

Fixes #97