extension: Drop the old schema file and migration code
In (815716b931: "extension: Rename gschema file to avoid conflict with
other extensions"), we introduced a new schema file while preserving the
old one so that users could have time to migrate their customizations to
the new schema.

Users have had ~6 months to update, so it's time to drop the old file so
that we're no longer conflicting with the standard
gnome-shell-extension-system-monitor extension. Anyone who missed this
update window will lose their customizations. Not a huge deal, resolving
the filename conflict is the priority at this point.

This finalizes #98.