Makefile: Drop all use of sudo
If the user is installing to a protected directory they should just invoke
`make` with sudo. There's no need for us to make assumptions about which
commands will need elevated privileges, and the current conflation of a
non-empty DESTDIR with needing SUDO doesn't work well for distro packaging,
resulting in the BUILD_FOR_RPM sudo override.

Rather than requiring distro packagers to roll their own build steps for
the extension, or requiring that they use our build system and set the
BUILD_FOR_RPM flag (even if they're *not* building for RPM), just drop all
use of `sudo` and rely on the user to invoke `make install` with sudo if
needed.